[Decomp][Character] Cohesion split: extract customization/equipment handlers to CharacterHandlerCustomize.cpp (#276)
Re-applies the mangosthree decomposition onto MangosOne's own TBC 2.4.3
source. Pure code move out of the parent translation unit into the new
cohesively-named file(s); no behaviour change (source diff is deletion-only).
CMake file(GLOB) picks the new TU(s) up automatically. Verified: game.lib
compiles and links clean (MSVC, Release).

Not ported (absent in MangosOne TBC 2.4.3 -- these are
later-expansion methods that do not exist in this branch):
  WorldSession::HandleAlterAppearanceOpcode,WorldSession::HandleRemoveGlyphOpcode,WorldSession::HandleCharCustomizeOpcode,WorldSession::HandleEquipmentSetSaveOpcode,WorldSession::HandleEquipmentSetDeleteOpcode,WorldSession::HandleEquipmentSetUseOpcode,WorldSession::HandleReorderCharactersOpcode,WorldSession::HandleSetCurrencyFlagsOpcode,WorldSession::HandleLoadScreenOpcode
WorldSession::HandleAlterAppearanceOpcode,WorldSession::HandleRemoveGlyphOpcode,WorldSession::HandleCharCustomizeOpcode,WorldSession::HandleEquipmentSetSaveOpcode,WorldSession::HandleEquipmentSetDeleteOpcode,WorldSession::HandleEquipmentSetUseOpcode,WorldSession::HandleReorderCharactersOpcode,WorldSession::HandleSetCurrencyFlagsOpcode,WorldSession::HandleLoadScreenOpcode
Ported-from: mangosthree/server@a3b6d83b5201
