+++ src/game/WorldHandlers/CharacterHandlerCustomize.cpp
@@ -1,0 +1,262 @@
+/**
+ * MaNGOS is a full featured server for World of Warcraft, supporting
+ * the following clients: 1.12.x, 2.4.3, 3.3.5a, 4.3.4a and 5.4.8
+ *
+ * Copyright (C) 2005-2025 MaNGOS <https://www.getmangos.eu>
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ *
+ * World of Warcraft, and all World of Warcraft or Warcraft art, images,
+ * and lore are copyrighted by Blizzard Entertainment, Inc.
+ */
+
+/**
+ * @file CharacterHandlerCustomize.cpp
+ * @brief Cohesion split of CharacterHandler.cpp.
+ *        Re-applied onto MangosOne TBC 2.4.3; same class, pure code move,
+ *        no behaviour change. CMake file(GLOB) picks this TU up automatically.
+ */
+
+#include "Common.h"
+#include "Database/DatabaseEnv.h"
+#include "WorldPacket.h"
+#include "SharedDefines.h"
+#include "WorldSession.h"
+#include "Opcodes.h"
+#include "Log.h"
+#include "World.h"
+#include "ObjectMgr.h"
+#include "Player.h"
+#include "CinematicFlyover.h"
+#include "Guild.h"
+#include "GuildMgr.h"
+#include "UpdateMask.h"
+#include "Auth/md5.h"
+#include "ObjectAccessor.h"
+#include "Group.h"
+#include "Database/DatabaseImpl.h"
+#include "PlayerDump.h"
+#include "SocialMgr.h"
+#include "Util.h"
+#include "ArenaTeam.h"
+#include "Language.h"
+#include "Chat.h"
+#include "SpellMgr.h"
+#include "GameTime.h"
+#include "Timer.h"
+#ifdef ENABLE_ELUNA
+#include "LuaEngine.h"
+#endif /* ENABLE_ELUNA */
+#ifdef ENABLE_PLAYERBOTS
+#include "playerbot.h"
+#include "PlayerbotAIConfig.h"
+#endif
+
+// config option SkipCinematics supported values
+enum CinematicsSkipMode
+{
+    CINEMATICS_SKIP_NONE      = 0,
+    CINEMATICS_SKIP_SAME_RACE = 1,
+    CINEMATICS_SKIP_ALL       = 2
+};
+
+/**
+ * @brief Validates and starts the asynchronous character rename flow.
+ *
+ * @param recv_data The received opcode packet.
+ */
+void WorldSession::HandleCharRenameOpcode(WorldPacket& recv_data)
+{
+    ObjectGuid guid;
+    std::string newname;
+
+    recv_data >> guid;
+    recv_data >> newname;
+
+    // prevent character rename to invalid name
+    if (!normalizePlayerName(newname))
+    {
+        WorldPacket data(SMSG_CHAR_RENAME, 1);
+        data << uint8(CHAR_NAME_NO_NAME);
+        SendPacket(&data);
+        return;
+    }
+
+    uint8 res = ObjectMgr::CheckPlayerName(newname, true);
+    if (res != CHAR_NAME_SUCCESS)
+    {
+        WorldPacket data(SMSG_CHAR_RENAME, 1);
+        data << uint8(res);
+        SendPacket(&data);
+        return;
+    }
+
+    // check name limitations
+    if (GetSecurity() == SEC_PLAYER && sObjectMgr.IsReservedName(newname))
+    {
+        WorldPacket data(SMSG_CHAR_RENAME, 1);
+        data << uint8(CHAR_NAME_RESERVED);
+        SendPacket(&data);
+        return;
+    }
+
+    std::string escaped_newname = newname;
+    CharacterDatabase.escape_string(escaped_newname);
+
+    // make sure that the character belongs to the current account, that rename at login is enabled
+    // and that there is no character with the desired new name
+    CharacterDatabase.AsyncPQuery(&WorldSession::HandleChangePlayerNameOpcodeCallBack,
+                                  GetAccountId(), newname,
+                                  "SELECT `guid`, `name` FROM `characters` WHERE `guid` = %u AND `account` = %u AND (`at_login` & %u) = %u AND NOT EXISTS (SELECT NULL FROM `characters` WHERE `name` = '%s')",
+                                  guid.GetCounter(), GetAccountId(), AT_LOGIN_RENAME, AT_LOGIN_RENAME, escaped_newname.c_str()
+                                 );
+}
+
+/**
+ * @brief Finalizes a character rename after the database validation query completes.
+ *
+ * @param result The rename validation query result.
+ * @param accountId The session account id.
+ * @param newname The requested new character name.
+ */
+void WorldSession::HandleChangePlayerNameOpcodeCallBack(QueryResult* result, uint32 accountId, std::string newname)
+{
+    WorldSession* session = sWorld.FindSession(accountId);
+    if (!session)
+    {
+        delete result;
+        return;
+    }
+
+    if (!result)
+    {
+        WorldPacket data(SMSG_CHAR_RENAME, 1);
+        data << uint8(CHAR_CREATE_ERROR);
+        session->SendPacket(&data);
+        return;
+    }
+
+    uint32 guidLow = result->Fetch()[0].GetUInt32();
+    ObjectGuid guid = ObjectGuid(HIGHGUID_PLAYER, guidLow);
+    std::string oldname = result->Fetch()[1].GetCppString();
+
+    delete result;
+
+    CharacterDatabase.BeginTransaction();
+    CharacterDatabase.PExecute("UPDATE `characters` SET `name` = '%s', `at_login` = `at_login` & ~ %u WHERE `guid` ='%u'", newname.c_str(), uint32(AT_LOGIN_RENAME), guidLow);
+    CharacterDatabase.PExecute("DELETE FROM `character_declinedname` WHERE `guid` ='%u'", guidLow);
+    CharacterDatabase.CommitTransaction();
+
+    sLog.outChar("Account: %d (IP: %s) Character:[%s] (guid:%u) Changed name to: %s", session->GetAccountId(), session->GetRemoteAddress().c_str(), oldname.c_str(), guidLow, newname.c_str());
+
+    WorldPacket data(SMSG_CHAR_RENAME, 1 + 8 + (newname.size() + 1));
+    data << uint8(RESPONSE_SUCCESS);
+    data << guid;
+    data << newname;
+    session->SendPacket(&data);
+
+    sWorld.InvalidatePlayerDataToAllClient(guid);
+}
+
+void WorldSession::HandleSetPlayerDeclinedNamesOpcode(WorldPacket& recv_data)
+{
+    ObjectGuid guid;
+
+    recv_data >> guid;
+
+    // not accept declined names for unsupported languages
+    std::string name;
+    if (!sObjectMgr.GetPlayerNameByGUID(guid, name))
+    {
+        WorldPacket data(SMSG_SET_PLAYER_DECLINED_NAMES_RESULT, 4 + 8);
+        data << uint32(1);
+        data << ObjectGuid(guid);
+        SendPacket(&data);
+        return;
+    }
+
+    std::wstring wname;
+    if (!Utf8toWStr(name, wname))
+    {
+        WorldPacket data(SMSG_SET_PLAYER_DECLINED_NAMES_RESULT, 4 + 8);
+        data << uint32(1);
+        data << ObjectGuid(guid);
+        SendPacket(&data);
+        return;
+    }
+
+    if (!isCyrillicCharacter(wname[0]))                     // name already stored as only single alphabet using
+    {
+        WorldPacket data(SMSG_SET_PLAYER_DECLINED_NAMES_RESULT, 4 + 8);
+        data << uint32(1);
+        data << ObjectGuid(guid);
+        SendPacket(&data);
+        return;
+    }
+
+    std::string name2;
+    DeclinedName declinedname;
+
+    recv_data >> name2;
+
+    if (name2 != name)                                      // character have different name
+    {
+        WorldPacket data(SMSG_SET_PLAYER_DECLINED_NAMES_RESULT, 4 + 8);
+        data << uint32(1);
+        data << ObjectGuid(guid);
+        SendPacket(&data);
+        return;
+    }
+
+    for (int i = 0; i < MAX_DECLINED_NAME_CASES; ++i)
+    {
+        recv_data >> declinedname.name[i];
+        if (!normalizePlayerName(declinedname.name[i]))
+        {
+            WorldPacket data(SMSG_SET_PLAYER_DECLINED_NAMES_RESULT, 4 + 8);
+            data << uint32(1);
+            data << ObjectGuid(guid);
+            SendPacket(&data);
+            return;
+        }
+    }
+
+    if (!ObjectMgr::CheckDeclinedNames(GetMainPartOfName(wname, 0), declinedname))
+    {
+        WorldPacket data(SMSG_SET_PLAYER_DECLINED_NAMES_RESULT, 4 + 8);
+        data << uint32(1);
+        data << ObjectGuid(guid);
+        SendPacket(&data);
+        return;
+    }
+
+    for (int i = 0; i < MAX_DECLINED_NAME_CASES; ++i)
+    {
+        CharacterDatabase.escape_string(declinedname.name[i]);
+    }
+
+    CharacterDatabase.BeginTransaction();
+    CharacterDatabase.PExecute("DELETE FROM `character_declinedname` WHERE `guid` = '%u'", guid.GetCounter());
+    CharacterDatabase.PExecute("INSERT INTO `character_declinedname` (`guid`, `genitive`, `dative`, `accusative`, `instrumental`, `prepositional`) VALUES ('%u','%s','%s','%s','%s','%s')",
+                               guid.GetCounter(), declinedname.name[0].c_str(), declinedname.name[1].c_str(), declinedname.name[2].c_str(), declinedname.name[3].c_str(), declinedname.name[4].c_str());
+    CharacterDatabase.CommitTransaction();
+
+    WorldPacket data(SMSG_SET_PLAYER_DECLINED_NAMES_RESULT, 4 + 8);
+    data << uint32(0);                                      // OK
+    data << ObjectGuid(guid);
+    SendPacket(&data);
+
+    sWorld.InvalidatePlayerDataToAllClient(guid);
+}
